Working singleton with on demand context
The singleton works, but in order to call get reference you now need a context, which I don't like but this is literally the only way we can do this with how Android is configured. See (https://developer.android.com/reference/android/app/Application) for details on why that is. The app runs with this commit, but there are a few bugs to work out and the full functionality of this issue has not been fixed yet.