Started making communication framework
NON-FUNCTIONAL COMMIT. This code is a starting point for sending and receiving data. As far as I can tell, it is one way, which is something that I need to change. The process of writing and saving data still needs to be established. and there is currently no calls to the code that is actually transmitting data.