Attempted to derive two-way communication.
I tried to make two way communication work, but there is a missing piece of information that I don't know how to derive. The client is aware of the host as a INetAddress, but the same can't be said for a Host about its Clients. Without that information, we can't transmit data.