reconfigured WifiDirectBroadcastReceiver
WifiDirectBroadcastReceiver now makes function calls which notify the WifiDirect singleton when group information and connection information is available, which was something that I assumed that was already occuring, but realized I was never subscribing. In addition, I refactored the file structure a tiny bit to reflect the nature of certain classes.