decoupled asynchronous tasks
I decoupled asynchronous tasks from the WifiDirect class by using the observer pattern judiciously. Really just went at it with callbacks and interfaces and so on. This actually simplied the code in the WifiDirect class a little bit, removed context leak issues. and also actually calls the functions that start asynchronous tasks. So, everything should work at this point, with maybe one or two issues that we shall see.